feat: web config 추가

+++ src/main/java/capstone/TryAngle/repository/ChallengeRepository.java
@@ -3,6 +3,6 @@ package capstone.TryAngle.repository;
 import capstone.TryAngle.model.challenge.Challenge;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ChallengeRepository extends JpaRepository<Challenge, Long> {
+public interface ChallengeRepository extends JpaRepository<Challenge, Integer> {
 
 }
